# src/test/resources/openapi/collectra-api-v1-baseline.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Collectra API",
    "description" : "Tenant-aware collection and communications platform",
    "version" : "v1"
  },
  "paths" : {
...
    "/api/v1/contracts/{id}/activate" : {
      "post" : {
        "tags" : [ "contract-controller" ],
        "operationId" : "activate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContractResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "list_11",
        "parameters" : [ {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "channel",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "scheduledFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "scheduledTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "createdFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "createdTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt,desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseCampaignItem"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "create_12",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCampaignRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaignId}/runs" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "runs",
        "parameters" : [ {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseRunItem"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "prepare",
        "parameters" : [ {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PrepareResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaignId}/activate" : {
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "activate_1",
        "parameters" : [ {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/runs/{runId}/eligibility-recheck" : {
      "post" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "recheck",
        "parameters" : [ {
          "name" : "runId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EligibilityResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{campaignId}" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "get_9",
        "parameters" : [ {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CampaignResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaignId}/runs/{runId}/recipients" : {
      "get" : {
        "tags" : [ "campaign-controller" ],
        "operationId" : "recipients",
        "parameters" : [ {
          "name" : "campaignId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "runId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "channel",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "customerId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseRecipientItem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "JsonNode" : { },
...
      "ContractResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "externalId" : {
            "type" : "string"
          },
          "contractNumber" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "SUSPENDED", "CLOSED", "CANCELLED" ]
          },
          "validFrom" : {
            "type" : "string",
            "format" : "date"
          },
          "validTo" : {
            "type" : "string",
            "format" : "date"
          },
          "renewalDate" : {
            "type" : "string",
            "format" : "date"
          },
          "customFields" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "VersionRequest" : {
        "type" : "object",
        "properties" : {
          "locale" : {
            "type" : "string",
            "maxLength" : 35,
            "minLength" : 0
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "EMAIL", "SMS", "WHATSAPP", "TELEGRAM", "PDF" ]
          },
          "subject" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "contentHtml" : {
            "type" : "string"
          },
          "stylesheet" : {
            "type" : "string"
          }
        },
        "required" : [ "locale" ]
      },
...
      "CampaignSelection" : {
        "type" : "object",
        "properties" : {
          "customerIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "uniqueItems" : true
          },
          "segmentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "uniqueItems" : true
          },
          "daysOverdueFrom" : {
            "type" : "integer",
            "format" : "int32"
          },
          "daysOverdueTo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "amountFrom" : {
            "type" : "string",
            "pattern" : "^-?(0|[1-9][0-9]{0,14})(\\.[0-9]{1,4})?$"
          },
          "amountTo" : {
            "type" : "string",
            "pattern" : "^-?(0|[1-9][0-9]{0,14})(\\.[0-9]{1,4})?$"
          }
        }
      },
      "CreateCampaignRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          },
          "templateVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "channel" : {
            "type" : "string",
            "maxLength" : 30,
            "minLength" : 0
          },
          "scheduledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "selection" : {
            "$ref" : "#/components/schemas/CampaignSelection"
          }
        },
        "required" : [ "channel", "name", "templateVersionId" ]
      },
      "CampaignResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "DRAFT", "ACTIVE", "ARCHIVED" ]
          },
          "templateVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "channel" : {
            "type" : "string"
          },
          "scheduledAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PrepareResult" : {
        "type" : "object",
        "properties" : {
          "runId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "recipients" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "EligibilityResult" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int32"
          },
          "eligible" : {
            "type" : "integer",
            "format" : "int32"
          },
          "skipped" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "CampaignItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "templateVersionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "channel" : {
            "type" : "string"
          },
          "scheduledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "PageResponseCampaignItem" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CampaignItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      },
      "PageResponseRunItem" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RunItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      },
      "RunItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "campaignId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string"
          },
          "recipientCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sentCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "failedCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "skippedCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pendingCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "preparedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "PageResponseRecipientItem" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RecipientItem"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      },
      "RecipientItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "invoiceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "channel" : {
            "type" : "string"
          },
          "destination" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "skipReason" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}